Implement music
The music starts muted. The music files/playlist is associated with the
theme. If a new theme is selected, the current music files will be
updated as well. Every time, the theme is updated, it will check if the
new music should be played or not depending on if the audio button is
unmuted(has textDecoration: line-through).